+++ src/main/java/br/com/escola/client/http/controller/ProfessorController.java
@@ -2,23 +2,24 @@ package br.com.escola.client.http.controller;
 
 
 import br.com.escola.client.entity.Professor;
-import br.com.escola.client.service.ProfTurmaService;
 import br.com.escola.client.service.ProfessorService;
+import org.hibernate.annotations.Parameter;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/professor")
+
 public class ProfessorController {
 
     @Autowired
     ProfessorService professorService;
-
 
     @Autowired
     ModelMapper modelMapper;
@@ -28,73 +29,54 @@ public class ProfessorController {
     ////////////////////////////////////////////////////////////////////
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public Professor saveProfessor(@RequestBody Professor professor){
-
-
-
+    public Professor saveProfessor(@RequestBody Professor professor) {
         return professorService.save(professor);
     }
 
 
-    ///////////////////////////////////GET ALL
+    ///////////////////////////////////GET
     ////////////////////////////////////////////////////////////////////
     @GetMapping
-    @ResponseStatus(HttpStatus.OK)
-    public List<Professor> getProfessor(){
-        return professorService.getProfessor();
+    public ResponseEntity getProfessor(@RequestParam(required = false, name = "cpf") Optional<String> cpf) {
+
+
+        if (cpf.isEmpty()) {
+            var found = professorService.getProfessor();
+            if (found.isEmpty()) return new ResponseEntity(HttpStatus.NOT_FOUND);
+            return new ResponseEntity(found, HttpStatus.OK);
+        }
+
+        var found = professorService.findByCpf(cpf.get());
+        if (found.isEmpty()) return new ResponseEntity(HttpStatus.NOT_FOUND);
+        return new ResponseEntity(found, HttpStatus.OK);
     }
-
-
-
-    ///////////////////////////////////GET BY CPF
-    ////////////////////////////////////////////////////////////////////
-    @GetMapping("/search/{cpf}")
-    @ResponseStatus(HttpStatus.OK)
-    public Professor findProfessor(@PathVariable("cpf") String cpf){
-
-        return professorService.findByCpf(cpf);
-
-    }
-
-
 
 
     ///////////////////////////////////DELETE BY CPF
     ////////////////////////////////////////////////////////////////////
-    @DeleteMapping("/{cpf}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void DeleteById(@PathVariable("cpf") String cpf){
+    @DeleteMapping
+    public ResponseEntity deleteById(@RequestParam("cpf") String cpf) {
+        var p = professorService.findByCpf(cpf);
+        if (p.isEmpty()) return new ResponseEntity(HttpStatus.NOT_FOUND);
+        var id = p.get().getId();
+
+
+        professorService.deleteDependency(id);
         professorService.deleteProfessorByCpf(cpf);
+        return new ResponseEntity(HttpStatus.OK);
     }
-
-
 
 
     ///////////////////////////////////MODIFY BY CPF
     ////////////////////////////////////////////////////////////////////
-    @PutMapping("/{cpf}/{element}={value}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public Professor updateProfessor(@PathVariable("cpf") String cpf, @PathVariable("element") String element,
-                                @PathVariable("value") String value){
-        Professor p = professorService.findByCpf(cpf);
-        switch(element){
-            case "cpf":
-                p.setCpf(value);
-                break;
-
-            case "nome":
-                p.setNome(value);
-                break;
-
-            case "email":
-                p.setEmail(value);
-                break;
-
-        }
-
-            return professorService.save(p);
-        }
-
+    @PutMapping
+    public ResponseEntity updateProfessor(@RequestParam("cpf") String cpf, @RequestBody Professor incomingBody) {
+        var p = professorService.findByCpf(cpf);
+        if (p.isEmpty()) return new ResponseEntity(HttpStatus.NOT_FOUND);
+        modelMapper.map(incomingBody, p.get());
+        professorService.save(p.get());
+        return new ResponseEntity(HttpStatus.OK);
+    }
 
 
 }
